# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "승뮤니티 API 문서화"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/register": {
      "get": {
        "tags": ["User"],
        "summary": "회원가입 페이지 조회",
        "description": "회원가입 페이지를 조회합니다.",
        "responses": {
          "200": {
            "description": "회원가입 페이지 조회 성공"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "사용자 등록",
        "description": "새로운 사용자를 회원가입 시킵니다.",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string",
            "maxLength": 20,
            "description": "등록할 사용자의 아이디"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "등록할 사용자의 비밀번호"
          }
        ],
        "responses": {
          "200": {
            "description": "회원가입 성공"
          },
          "400": {
            "description": "아이디 또는 비밀번호가 20글자를 초과하는 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('아이디 또는 비밀번호는 20글자 이하로 지정해주세요.');window.location.replace(location.href);</script>"
            }
          },
          "409": {
            "description": "같은 아이디가 이미 존재하는 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('이미 존재하는 아이디입니다');location.replace(location.href);</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": ["User"],
        "summary": "로그인 페이지 조회",
        "description": "로그인 페이지를 조회합니다.",
        "responses": {
          "200": {
            "description": "로그인 페이지 조회 성공"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "사용자 로그인",
        "description": "사용자를 인증하고 세션을 생성하여 로그인합니다.",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "사용자의 아이디"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": "사용자의 비밀번호"
          }
        ],
        "responses": {
          "200": {
            "description": "로그인 성공"
          },
          "401": {
            "description": "잘못된 사용자 아이디 또는 비밀번호",
            "schema": {
              "type": "string",
              "example": "<script>alert('인증 정보가 유효하지 않습니다.');history.go(-1)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": ["User"],
        "summary": "사용자 로그아웃",
        "description": "클라이언트의 세션을 파괴하고 사용자를 로그아웃 시킵니다.",
        "responses": {
          "302": {
            "description": "리다이렉트 성공"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/mypage": {
      "get": {
        "tags": ["User"],
        "summary": "마이페이지 조회",
        "description": "사용자의 마이페이지를 조회합니다.",
        "responses": {
          "200": {
            "description": "마이페이지 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "아이디": {
                  "$ref": "#/definitions/User"
                },
                "isRead": {
                  "type": "boolean",
                  "description": "읽음 여부"
                }
              }
            }
          },
          "401": {
            "description": "사용자가 로그인하지 않음",
            "schema": {
              "type": "string",
              "example": "<script>alert('로그인 요망');window.location.replace(`/login`)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/mypage/name": {
      "put": {
        "tags": ["User"],
        "summary": "닉네임 변경",
        "description": "사용자의 닉네임을 변경합니다.",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "description": "새로운 닉네임",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "new_username"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "닉네임 변경 성공"
          },
          "400": {
            "description": "아이디가 20글자를 초과하는 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('아이디는 20글자 이하로 지정해주세요.');window.location.replace(`/mypage`);</script>"
            }
          },
          "409": {
            "description": "이미 존재하는 아이디",
            "schema": {
              "type": "string",
              "example": "<script>alert('이미 존재하는 아이디 입니다');window.location.replace(`/mypage`);</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/mypage/profileImg": {
      "put": {
        "tags": ["User"],
        "summary": "프로필 사진 변경",
        "description": "사용자의 프로필 사진을 변경합니다.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "img1",
            "in": "formData",
            "description": "프로필 사진 파일",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 사진 변경 성공",
            "schema": {
              "type": "string",
              "example": "<script>alert('프로필 사진이 변경되었습니다');window.location.replace(`/mypage`);</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/locations": {
      "put": {
        "tags": ["User"],
        "summary": "사용자 위치 정보 업데이트",
        "description": "사용자의 위치 정보를 업데이트합니다.",
        "parameters": [
          {
            "name": "content",
            "in": "body",
            "description": "업데이트할 위치 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "위치 정보 업데이트 성공"
          },
          "400": {
            "description": "위치 정보가 제공되지 않음"
          },
          "401": {
            "description": "사용자가 로그인하지 않음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },   
    "/list": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 1페이지 조회",
        "description": "게시글의 1페이지를 조회합니다.",
        "responses": {
          "200": {
            "description": "게시글 1페이지 조회 성공"
          }
        }
      },
      "delete": {
        "tags": ["Post"],
        "summary": "특정 게시글 삭제",
        "description": "ID 값을 사용하여 특정 게시글을 삭제합니다.",
        "parameters": [
          {
            "name": "docid",
            "in": "query",
            "description": "게시물의 ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 삭제 완료"
          },
          "401": {
            "description": "사용자가 로그인하지 않음"
          },
          "403": {
            "description": "게시글을 삭제할 권한이 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
...
    "/write": {
      "get": {
        "tags": ["Post"],
        "summary": "게시글 작성 페이지 조회",
        "description": "게시글을 작성하는 페이지를 조회합니다.",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "글쓰기 페이지 조회 성공"
          },
          "401": {
            "description": "사용자가 로그인하지 않음",
            "schema": {
              "type": "string",
              "example": "<script>alert('로그인 요망');window.location.replace(`/login`)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      },
      "post": {
        "tags": ["Post"],
        "summary": "게시글 작성",
        "description": "사용자가 새로운 게시물을 작성합니다.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "formData",
            "name": "title",
            "description": "게시물 제목",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "content",
            "description": "게시물 내용",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "img1",
            "description": "이미지 파일",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 작성 성공"
          },
          "400": {
            "description": "제목이나 내용이 없는 경우 또는 제목이 20글자를 초과하는 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('제목을 20글자 이하로 지정해주세요.');history.go(-1)</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/edit": {
      "put": {
        "tags": ["Post"],
        "summary": "게시글 수정",
        "description": "사용자가 게시글을 수정합니다.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "수정할 게시물 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "수정할 게시물의 ID"
                },
                "userId": {
                  "type": "string",
                  "description": "게시물을 작성한 사용자의 ID"
                },
                "title": {
                  "type": "string",
                  "description": "게시물 제목 (20글자 이하)"
                },
                "content": {
                  "type": "string",
                  "description": "게시물 내용"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 수정 성공"
          },
          "400": {
            "description": "게시글 제목이 20글자를 초과하는 경우",
            "schema": {
              "type": "string",
              "example": "<script>alert('제목을 20글자 이하로 지정해주세요.');history.go(-1)</script>"
            }
          },
          "403": {
            "description": "게시글을 수정할 권한이 없음",
            "schema": {
              "type": "string",
              "example": "<script>alert('수정할 수 없습니다.');window.location.replace('/list/1')</script>"
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
...
    "/chat": {
      "get": {
        "tags": ["Chat"],
        "summary": "채팅방 조회",
        "description": "사용자가 속한 모든 채팅방을 조회합니다.",
        "responses": {
          "200": {
            "description": "채팅방 조회 성공",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                    }
                  }
                },
                "cur": {
                  "type": "string",
                  "description": "현재 사용자의 ID"
                },
                "arrow": {
                  "type": "integer",
                  "description": "화살표 변수"
                },
                "counterpart": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "상대방 이름 목록"
                }
              }
            }
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      },
      "delete": {
        "tags": ["Chat"],
        "summary": "채팅방 삭제",
        "description": "채팅방의 ID로 특정 채팅방을 삭제합니다.",
        "parameters": [
          {
            "name": "삭제id",
            "in": "body",
            "description": "삭제할 채팅방의 ID",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "삭제id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "채팅방 삭제 성공"
          },
          "400": {
            "description": "삭제할 채팅방이 존재하지 않음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
...
  },
...
}